In this commit i finalised the sign in page functionality including the error handling. All data from sign up page is inserted into the Users table in the database.

+++ app/src/main/java/data/User.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 //primary constructor
 @Entity(tableName = "users")
 data class User(
-    @PrimaryKey(autoGenerate = true) val user_id : Long=0,
+    @PrimaryKey(autoGenerate = true) val user_id : Int =0,
     var username : String,
     var password :String,
     var email : String,
